Uses two decks of cards, introduces fifth suit.
With 104 cards, I intend to use:
* 4 cards as player crib sheets
* 5 cards to represent the remaining number of rounds.
* 5 suits, each with 19 cards. For each suit, there will be 5 campaign
  starter cards and 14 event cards.

This change adds a fifth suit, the "debate", which is a risky suit but
doesn't favour stealing or trashing. The existing suits have had more
cards added to them, which has kept the ratios roughly the same. There's
some slight differences tending towards negative, with a lower percentage
of permutations that score zero overall!